Added more weapon options to game config.

+++ MatrixGame/src/Effects/MatrixEffectMovingObject.cpp
@@ -209,7 +209,7 @@ void MO_Cannon_Round_Tact(D3DXMATRIX& m, SMOProps& props, float tact)
 
     if(props.gun.in_frustum_count > 0 && props.gun.contrail_sprite_num != SPR_NONE)
     {
-        CMatrixEffect::CreateSpritesLine(nullptr, props.curpos, hitpos, (float)props.gun.contrail_width, props.gun.contrail_color, 0, (float)props.gun.contrail_duration, CMatrixEffect::GetSingleBrightSpriteTex(props.gun.contrail_sprite_num));
+        CMatrixEffect::CreateSpritesLine(nullptr, props.curpos, hitpos, props.gun.contrail_width, props.gun.contrail_color, 0, props.gun.contrail_duration, CMatrixEffect::GetSingleBrightSpriteTex(props.gun.contrail_sprite_num));
         //DM("cc", CStr(props.curpos.x) + "," + CStr(props.curpos.y) + "," + CStr(props.curpos.z) + "," + CStr(hitpos.x) + "," + CStr(hitpos.x) + "," + CStr(hitpos.x));
     }
 
@@ -225,8 +225,8 @@ void MO_Cannon_Round_Tact(D3DXMATRIX& m, SMOProps& props, float tact)
             bool fire = false;
             if(hito == TRACE_STOP_LANDSCAPE)
             {
-                CMatrixEffect::CreateLandscapeSpot(nullptr, D3DXVECTOR2(hitpos.x, hitpos.y), FSRND((float)M_PI), FRND(3) + 6, SPOT_VORONKA);
-                hitpos.z = g_MatrixMap->GetZ(hitpos.x, hitpos.y) + 10;
+                CMatrixEffect::CreateLandscapeSpot(nullptr, D3DXVECTOR2(hitpos.x, hitpos.y), FSRND((float)M_PI), FRND(3.0f) + 6.0f, SPOT_VORONKA);
+                hitpos.z = g_MatrixMap->GetZ(hitpos.x, hitpos.y) + 10.0f;
                 fire = true;
             }
             CMatrixEffect::CreateExplosion(hitpos, ExplosionMissile, fire);
