+++ MatrixGame/src/MatrixTransition.cpp
@@ -79,12 +79,13 @@ void CTransition::BuildTexture()
         coord_mod = double(m_ScreenY) / m_ScreenX;
         if(coord_mod != 0.75) //0.75 - базовое игровое соотношение сторон 4 на 3
         {
-            x_offset = (m_ScreenX - m_ScreenY / 0.75) / 2;
-            native_aspect_res = m_ScreenX - (m_ScreenX - m_ScreenY / 0.75);
+            x_offset = int(m_ScreenX - m_ScreenY / 0.75) / 2;
+            native_aspect_res = m_ScreenX - int(m_ScreenX - m_ScreenY / 0.75);
         }
 
-        coord_mod = m_ScreenY / 768.0;
+        coord_mod = double(m_ScreenY) / 768.0;
     }
+
 
     //≈сли было выставлено (не знаю как и зачем) кастомное разрешение не равное ни 1024x768, ни 800x600
     //Ћибо если поломан конфиг
@@ -114,42 +115,43 @@ void CTransition::BuildTexture()
         const CWStr& da = tra->ParGet(i);
 
         int x_check = da.GetStrPar(0, L"|").GetIntPar(0, L",");
+
         if(!x_check) m_Geom[i].verts[0].p.x = 0;
         else if(x_check == 1024) m_Geom[i].verts[0].p.x = float(m_ScreenX);
-        else m_Geom[i].verts[0].p.x = da.GetStrPar(0, L"|").GetDoublePar(0, L",") * coord_mod + x_offset;
-        m_Geom[i].verts[0].p.y = da.GetStrPar(0, L"|").GetDoublePar(1, L",") * coord_mod;
-        m_Geom[i].verts[0].p.z = 0;
-        m_Geom[i].verts[0].p.w = 1;
+        else m_Geom[i].verts[0].p.x = float(da.GetStrPar(0, L"|").GetDoublePar(0, L",") * coord_mod + x_offset);
+        m_Geom[i].verts[0].p.y = float(da.GetStrPar(0, L"|").GetDoublePar(1, L",") * coord_mod);
+        m_Geom[i].verts[0].p.z = 0.0f;
+        m_Geom[i].verts[0].p.w = 1.0f;
         m_Geom[i].verts[0].tu = m_Geom[i].verts[0].p.x / float(bmout.SizeX());
         m_Geom[i].verts[0].tv = m_Geom[i].verts[0].p.y / float(bmout.SizeY());
 
         x_check = da.GetStrPar(1, L"|").GetIntPar(0, L",");
         if(!x_check) m_Geom[i].verts[1].p.x = 0;
         else if(x_check == 1024) m_Geom[i].verts[1].p.x = float(m_ScreenX);
-        else m_Geom[i].verts[1].p.x = da.GetStrPar(1, L"|").GetDoublePar(0, L",") * coord_mod + x_offset;
-        m_Geom[i].verts[1].p.y = da.GetStrPar(1, L"|").GetDoublePar(1, L",") * coord_mod;
-        m_Geom[i].verts[1].p.z = 0;
-        m_Geom[i].verts[1].p.w = 1;
+        else m_Geom[i].verts[1].p.x = float(da.GetStrPar(1, L"|").GetDoublePar(0, L",") * coord_mod + x_offset);
+        m_Geom[i].verts[1].p.y = float(da.GetStrPar(1, L"|").GetDoublePar(1, L",") * coord_mod);
+        m_Geom[i].verts[1].p.z = 0.0f;
+        m_Geom[i].verts[1].p.w = 1.0f;
         m_Geom[i].verts[1].tu = m_Geom[i].verts[1].p.x / float(bmout.SizeX());
         m_Geom[i].verts[1].tv = m_Geom[i].verts[1].p.y / float(bmout.SizeY());
 
         x_check = da.GetStrPar(3, L"|").GetIntPar(0, L",");
         if(!x_check) m_Geom[i].verts[2].p.x = 0;
         else if(x_check == 1024) m_Geom[i].verts[2].p.x = float(m_ScreenX);
-        else m_Geom[i].verts[2].p.x = da.GetStrPar(3, L"|").GetDoublePar(0, L",") * coord_mod + x_offset;
-        m_Geom[i].verts[2].p.y = da.GetStrPar(3, L"|").GetDoublePar(1, L",") * coord_mod;
-        m_Geom[i].verts[2].p.z = 0;
-        m_Geom[i].verts[2].p.w = 1;
+        else m_Geom[i].verts[2].p.x = float(da.GetStrPar(3, L"|").GetDoublePar(0, L",") * coord_mod + x_offset);
+        m_Geom[i].verts[2].p.y = float(da.GetStrPar(3, L"|").GetDoublePar(1, L",") * coord_mod);
+        m_Geom[i].verts[2].p.z = 0.0f;
+        m_Geom[i].verts[2].p.w = 1.0f;
         m_Geom[i].verts[2].tu = m_Geom[i].verts[2].p.x / float(bmout.SizeX());
         m_Geom[i].verts[2].tv = m_Geom[i].verts[2].p.y / float(bmout.SizeY());
 
         x_check = da.GetStrPar(2, L"|").GetIntPar(0, L",");
         if(!x_check) m_Geom[i].verts[3].p.x = 0;
         else if(x_check == 1024) m_Geom[i].verts[3].p.x = float(m_ScreenX);
-        else m_Geom[i].verts[3].p.x = da.GetStrPar(2, L"|").GetDoublePar(0, L",") * coord_mod + x_offset;
-        m_Geom[i].verts[3].p.y = da.GetStrPar(2, L"|").GetDoublePar(1, L",") * coord_mod;
-        m_Geom[i].verts[3].p.z = 0;
-        m_Geom[i].verts[3].p.w = 1;
+        else m_Geom[i].verts[3].p.x = float(da.GetStrPar(2, L"|").GetDoublePar(0, L",") * coord_mod + x_offset);
+        m_Geom[i].verts[3].p.y = float(da.GetStrPar(2, L"|").GetDoublePar(1, L",") * coord_mod);
+        m_Geom[i].verts[3].p.z = 0.0f;
+        m_Geom[i].verts[3].p.w = 1.0f;
         m_Geom[i].verts[3].tu = m_Geom[i].verts[3].p.x / float(bmout.SizeX());
         m_Geom[i].verts[3].tv = m_Geom[i].verts[3].p.y / float(bmout.SizeY());
